Exported the displayCurrentTime
